refactor: Move default callbacks to util.h
