+++ src/secp256k1.c
@@ -44,34 +44,6 @@
     } \
 } while(0)
 
-#ifndef USE_EXTERNAL_DEFAULT_CALLBACKS
-#include <stdlib.h>
-#include <stdio.h>
-static void secp256k1_default_illegal_callback_fn(const char* str, void* data) {
-    (void)data;
-    fprintf(stderr, "[libsecp256k1] illegal argument: %s\n", str);
-    abort();
-}
-static void secp256k1_default_error_callback_fn(const char* str, void* data) {
-    (void)data;
-    fprintf(stderr, "[libsecp256k1] internal consistency check failed: %s\n", str);
-    abort();
-}
-#else
-void secp256k1_default_illegal_callback_fn(const char* str, void* data);
-void secp256k1_default_error_callback_fn(const char* str, void* data);
-#endif
-
-static const secp256k1_callback default_illegal_callback = {
-    secp256k1_default_illegal_callback_fn,
-    NULL
-};
-
-static const secp256k1_callback default_error_callback = {
-    secp256k1_default_error_callback_fn,
-    NULL
-};
-
 struct secp256k1_context_struct {
     secp256k1_ecmult_gen_context ecmult_gen_ctx;
     secp256k1_callback illegal_callback;
